Remove broken `cabal sandbox add-source` call.
The add-source subcommand expects a directory name, which should contain
a *.cabal file. A path to a *.cabal file does not meet those
requirements. Further, cabal already searches the working directory for
source code, so the call was unnecessary.

This bug made the build fail the first time `make` was run in a clean
checkout, but since the `cabal sandbox init` call has the side effect of
creating .cabal-sandbox, make didn't try to run this rule on the second
and subsequent tries.